split up the inventory process
