+++ src/api/java/baritone/api/Settings.java
@@ -500,6 +500,18 @@ public final class Settings {
     public final Setting<Boolean> containerMemory = new Setting<>(false);
 
     /**
+     * This stores excess inventory- it will try to keep one slot in your inventory open
+     * This will store items in shulker chests as well as nearby inventories
+     */
+    public final Setting<Boolean> storeExcessInventory = new Setting<>(false);
+
+    /**
+     * Stores excess inventory in remembered chests/inventories
+     * You must have storeExcessInventory on for this to work
+     */
+    public final Setting<Boolean> storeExcessInChests = new Setting<>(false);
+    
+    /**
      * Fill in blocks behind you
      */
     public final Setting<Boolean> backfill = new Setting<>(false);
